% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchpipe.R
\name{branch}
\alias{branch}
\title{branch and merge}
\usage{
branch(.data, branch_name, ...)
}
\value{
An object of the same time as .data, as with all pipe verbs.
}
\description{
Create more complex data flows in a series
of magrittr/dplyr pipes.
}
\examples{
mtcars \%>\%
    filter(wt > 2) \%>\%
    branch('common_gearings') \%>\%
        group_by(gear) \%>\%
        count() \%>\%
        filter(n > 8) \%>\%
    merge() \%>\%
    left_join(x = ., y = common_gearings)

}
